third_party/blink/web_tests/ppapi/plugins/embed-attributes-setting-expected.txt

CONSOLE MESSAGE: Blink Test Plugin: initializing
CONSOLE MESSAGE: Blink Test Plugin: initializing
[Embed is element specified in markup]
PASS: embed.getAttribute('align') should be 1 and is.
PASS: embed.getAttribute('height') should be 1 and is.
PASS: embed.getAttribute('name') should be 1 and is.
PASS: embed.getAttribute('width') should be 1 and is.
PASS: embed.getAttribute('type') should be 1 and is.
PASS: typeof embed.postMessage should be function and is.
----------
[Embed is dynamically created element with only type specified]
PASS: typeof embed.postMessage should be function and is.
 
